Move project with resources test to api level
