+++ tests/api/api/test_projects.py
@@ -13,9 +13,22 @@ from sqlalchemy.orm import Session
 
 import mlrun.api.crud
 import mlrun.api.schemas
+import mlrun.api.utils.singletons.db
 import mlrun.artifacts.dataset
 import mlrun.artifacts.model
 import mlrun.errors
+from mlrun.api.db.sqldb.models import (
+    Artifact,
+    Entity,
+    Feature,
+    FeatureSet,
+    FeatureVector,
+    Function,
+    Project,
+    Run,
+    Schedule,
+    _classes,
+)
 
 
 def test_create_project_failure_already_exists(db: Session, client: TestClient) -> None:
@@ -32,6 +45,60 @@ def test_create_project_failure_already_exists(db: Session, client: TestClient) 
     # create again
     response = client.post("/api/projects", json=project_1.dict())
     assert response.status_code == HTTPStatus.CONFLICT.value
+
+
+def test_delete_project_with_resources(db: Session, client: TestClient):
+    project_to_keep = "project-to-keep"
+    project_to_remove = "project-to-remove"
+    _create_resources_of_all_kinds(db, project_to_keep)
+    _create_resources_of_all_kinds(db, project_to_remove)
+    project_to_keep_table_name_records_count_map_before_project_removal = _assert_resources_in_project(
+        db, project_to_keep
+    )
+    _assert_resources_in_project(db, project_to_remove)
+
+    # deletion strategy - restricted - should fail because there are resources
+    response = client.delete(
+        f"/api/projects/{project_to_remove}",
+        headers={
+            mlrun.api.schemas.HeaderNames.deletion_strategy: mlrun.api.schemas.DeletionStrategy.restricted
+        },
+    )
+    assert response.status_code == HTTPStatus.PRECONDITION_FAILED.value
+
+    # mock runtime resources deletion
+    mlrun.api.crud.Runtimes().delete_runtimes = unittest.mock.Mock()
+
+    # deletion strategy - cascading - should succeed and remove all related resources
+    response = client.delete(
+        f"/api/projects/{project_to_remove}",
+        headers={
+            mlrun.api.schemas.HeaderNames.deletion_strategy: mlrun.api.schemas.DeletionStrategy.cascading
+        },
+    )
+    assert response.status_code == HTTPStatus.NO_CONTENT.value
+
+    project_to_keep_table_name_records_count_map_after_project_removal = _assert_resources_in_project(
+        db, project_to_keep
+    )
+    _assert_resources_in_project(db, project_to_remove, assert_no_resources=True)
+    assert (
+        deepdiff.DeepDiff(
+            project_to_keep_table_name_records_count_map_before_project_removal,
+            project_to_keep_table_name_records_count_map_after_project_removal,
+            ignore_order=True,
+        )
+        == {}
+    )
+
+    # deletion strategy - restricted - should succeed cause no project
+    response = client.delete(
+        f"/api/projects/{project_to_remove}",
+        headers={
+            mlrun.api.schemas.HeaderNames.deletion_strategy: mlrun.api.schemas.DeletionStrategy.restricted
+        },
+    )
+    assert response.status_code == HTTPStatus.NO_CONTENT.value
 
 
 def test_list_projects_summary_format(db: Session, client: TestClient) -> None:
@@ -279,6 +346,232 @@ def test_projects_crud(db: Session, client: TestClient) -> None:
 
     # list
     _list_project_names_and_assert(client, [name2])
+
+
+def _create_resources_of_all_kinds(db_session: Session, project: str):
+    db = mlrun.api.utils.singletons.db.get_db()
+    # add labels to project
+    project_schema = mlrun.api.schemas.Project(
+        metadata=mlrun.api.schemas.ProjectMetadata(
+            name=project, labels={"key": "value"}
+        ),
+        spec=mlrun.api.schemas.ProjectSpec(description="some desc"),
+    )
+    db.store_project(db_session, project, project_schema)
+
+    # Create several functions with several tags
+    labels = {
+        "name": "value",
+        "name2": "value2",
+    }
+    function = {
+        "bla": "blabla",
+        "metadata": {"labels": labels},
+        "status": {"bla": "blabla"},
+    }
+    function_names = ["function_name_1", "function_name_2", "function_name_3"]
+    function_tags = ["some_tag", "some_tag2", "some_tag3"]
+    for function_name in function_names:
+        for function_tag in function_tags:
+            db.store_function(
+                db_session,
+                function,
+                function_name,
+                project,
+                tag=function_tag,
+                versioned=True,
+            )
+
+    # Create several artifacts with several tags
+    artifact = {
+        "bla": "blabla",
+        "labels": labels,
+        "status": {"bla": "blabla"},
+    }
+    artifact_keys = ["artifact_key_1", "artifact_key_2", "artifact_key_3"]
+    artifact_uids = ["some_uid", "some_uid2", "some_uid3"]
+    artifact_tags = ["some_tag", "some_tag2", "some_tag3"]
+    for artifact_key in artifact_keys:
+        for artifact_uid in artifact_uids:
+            for artifact_tag in artifact_tags:
+                for artifact_iter in range(3):
+                    db.store_artifact(
+                        db_session,
+                        artifact_key,
+                        artifact,
+                        artifact_uid,
+                        artifact_iter,
+                        artifact_tag,
+                        project,
+                    )
+
+    # Create several runs
+    run = {
+        "bla": "blabla",
+        "metadata": {"labels": labels},
+        "status": {"bla": "blabla"},
+    }
+    run_uids = ["some_uid", "some_uid2", "some_uid3"]
+    for run_uid in run_uids:
+        for run_iter in range(3):
+            db.store_run(db_session, run, run_uid, project, run_iter)
+
+    # Create several logs
+    log = b"some random log"
+    log_uids = ["some_uid", "some_uid2", "some_uid3"]
+    for log_uid in log_uids:
+        db.store_log(db_session, log_uid, project, log)
+
+    # Create several schedule
+    schedule = {
+        "bla": "blabla",
+        "status": {"bla": "blabla"},
+    }
+    schedule_cron_trigger = mlrun.api.schemas.ScheduleCronTrigger(year=1999)
+    schedule_names = ["schedule_name_1", "schedule_name_2", "schedule_name_3"]
+    for schedule_name in schedule_names:
+        db.create_schedule(
+            db_session,
+            project,
+            schedule_name,
+            mlrun.api.schemas.ScheduleKinds.job,
+            schedule,
+            schedule_cron_trigger,
+            mlrun.config.config.httpdb.scheduling.default_concurrency_limit,
+            labels,
+        )
+
+    feature_set = mlrun.api.schemas.FeatureSet(
+        metadata=mlrun.api.schemas.ObjectMetadata(
+            name="dummy", tag="latest", labels={"owner": "nobody"}
+        ),
+        spec=mlrun.api.schemas.FeatureSetSpec(
+            entities=[
+                mlrun.api.schemas.Entity(
+                    name="ent1", value_type="str", labels={"label": "1"}
+                )
+            ],
+            features=[
+                mlrun.api.schemas.Feature(
+                    name="feat1", value_type="str", labels={"label": "1"}
+                )
+            ],
+        ),
+        status={},
+    )
+    db.create_feature_set(db_session, project, feature_set)
+
+    feature_vector = mlrun.api.schemas.FeatureVector(
+        metadata=mlrun.api.schemas.ObjectMetadata(
+            name="dummy", tag="latest", labels={"owner": "somebody"}
+        ),
+        spec=mlrun.api.schemas.ObjectSpec(),
+        status=mlrun.api.schemas.ObjectStatus(state="created"),
+    )
+    db.create_feature_vector(db_session, project, feature_vector)
+
+
+def _assert_resources_in_project(
+    db_session: Session, project: str, assert_no_resources: bool = False,
+) -> typing.Dict:
+    table_name_records_count_map = {}
+    for cls in _classes:
+        # User support is not really implemented or in use
+        # Run tags support is not really implemented or in use
+        if cls.__name__ != "User" and cls.__tablename__ != "runs_tags":
+            number_of_cls_records = 0
+            # Label doesn't have project attribute
+            # Project (obviously) doesn't have project attribute
+            # Features and Entities are not directly linked to project since they are sub-entity of feature-sets
+            if (
+                cls.__name__ != "Label"
+                and cls.__name__ != "Project"
+                and cls.__name__ != "Feature"
+                and cls.__name__ != "Entity"
+            ):
+                number_of_cls_records = (
+                    db_session.query(cls).filter_by(project=project).count()
+                )
+            elif cls.__name__ == "Label":
+                if cls.__tablename__ == "functions_labels":
+                    number_of_cls_records = (
+                        db_session.query(Function)
+                        .join(cls)
+                        .filter(Function.project == project)
+                        .count()
+                    )
+                if cls.__tablename__ == "runs_labels":
+                    number_of_cls_records = (
+                        db_session.query(Run)
+                        .join(cls)
+                        .filter(Run.project == project)
+                        .count()
+                    )
+                if cls.__tablename__ == "artifacts_labels":
+                    number_of_cls_records = (
+                        db_session.query(Artifact)
+                        .join(cls)
+                        .filter(Artifact.project == project)
+                        .count()
+                    )
+                if cls.__tablename__ == "feature_sets_labels":
+                    number_of_cls_records = (
+                        db_session.query(FeatureSet)
+                        .join(cls)
+                        .filter(FeatureSet.project == project)
+                        .count()
+                    )
+                if cls.__tablename__ == "features_labels":
+                    number_of_cls_records = (
+                        db_session.query(FeatureSet)
+                        .join(Feature)
+                        .join(cls)
+                        .filter(FeatureSet.project == project)
+                        .count()
+                    )
+                if cls.__tablename__ == "entities_labels":
+                    number_of_cls_records = (
+                        db_session.query(FeatureSet)
+                        .join(Entity)
+                        .join(cls)
+                        .filter(FeatureSet.project == project)
+                        .count()
+                    )
+                if cls.__tablename__ == "schedules_v2_labels":
+                    number_of_cls_records = (
+                        db_session.query(Schedule)
+                        .join(cls)
+                        .filter(Schedule.project == project)
+                        .count()
+                    )
+                if cls.__tablename__ == "feature_vectors_labels":
+                    number_of_cls_records = (
+                        db_session.query(FeatureVector)
+                        .join(cls)
+                        .filter(FeatureVector.project == project)
+                        .count()
+                    )
+                if cls.__tablename__ == "projects_labels":
+                    number_of_cls_records = (
+                        db_session.query(Project)
+                        .join(cls)
+                        .filter(Project.name == project)
+                        .count()
+                    )
+            else:
+                number_of_cls_records = (
+                    db_session.query(Project).filter(Project.name == project).count()
+                )
+            if assert_no_resources:
+                assert (
+                    number_of_cls_records == 0
+                ), f"Table {cls.__tablename__} records were found"
+            else:
+                assert (
+                    number_of_cls_records > 0
+                ), f"Table {cls.__tablename__} records were not found"
+            table_name_records_count_map[cls.__tablename__] = number_of_cls_records
+    return table_name_records_count_map
 
 
 def _list_project_names_and_assert(
